chore: cleanup helpers
