+++ apps/bot/src/handlers/reactionNotifications.ts
@@ -1,4 +1,4 @@
-import { Colors, EmbedBuilder, MessageReaction, User } from "discord.js";
+import { Client, Colors, EmbedBuilder, MessageReaction, User } from "discord.js";
 import { guildReactionNotifiers } from "..";
 
 export const handleReactionNotifications = async (reaction: MessageReaction, user: User) => {
@@ -31,4 +31,32 @@ export const handleReactionNotifications = async (reaction: MessageReaction, use
                 .catch(console.error)
         }
     }
+}
+
+
+export const updateCachedReactionNotifiers = async (
+    client: Client,
+    guilds: string[],
+    userId: User['id'],
+    targetId?: User['id'] | null
+) => {
+    const guildIds = guilds.length === 0 ?
+        //for all guilds
+        [
+            ...client.guilds.cache
+                .filter(async g => (
+                    //TODO?: use cached members
+                    await g.members.fetch()
+                ).has(userId))
+                .keys()
+        ] : guilds
+
+    for (const guildId of guildIds) {
+        const g = guildReactionNotifiers.get(guildId);
+        guildReactionNotifiers.set(guildId, {
+            users: [...g?.users.values() ?? [], userId],
+            targetId: targetId ?? g?.targetId
+        })
+    }
+
 }
